backend/requirements.txt
# BeatBridge Backend Requirements

# Flask and extensions
flask==2.0.1
flask-cors==3.0.10
gunicorn==20.1.0

# API Authentication
pyjwt==2.1.0
cryptography==35.0.0

# Environment variables
python-dotenv==0.19.0

# HTTP requests
requests==2.26.0

# Background tasks
redis==3.5.3
rq==1.10.0

# Spotify API
spotipy==2.19.0

# YouTube API
google-api-python-client==2.23.0
google-auth==2.2.1
google-auth-oauthlib==0.4.6
google-auth-httplib2==0.1.0

# Utilities
python-dateutil==2.8.2
fuzzywuzzy==0.18.0
python-Levenshtein==0.12.2